refactor: 변수명 변경후 바꾸지 못한 Mapper 모두 변경
