--- app/src/main/java/woowacourse/shopping/mapper/OrderListMapper.kt
@@ -1,14 +1,0 @@
-package woowacourse.shopping.mapper
-
-import woowacourse.shopping.model.Order
-import woowacourse.shopping.model.OrderUIModel
-
-fun Order.toUIModel(): OrderUIModel {
-    return OrderUIModel(
-        cartProducts = cartItems.map { it.toUIModel(true) },
-        totalPrice = totalPrice,
-        currentPoints = currentPoints,
-        earnedPoints = earnedPoints,
-        availablePoints = availablePoints
-    )
-}
